Added Identity server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -56,12 +57,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
-            // TODO: add current user as seller
-            auction.Seller = "user";
+
+            auction.Seller = User.Identity.Name;
             _context.Auctions.Add(auction);
 
             // Publish to Queue (before saving to DB for Atomicity)
@@ -76,6 +78,7 @@ namespace AuctionService.Controllers
             return CreatedAtAction(nameof(GetAuction), new { id = auction.Id }, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto auctionDto)
         {
@@ -84,7 +87,9 @@ namespace AuctionService.Controllers
                 .FirstOrDefaultAsync(x => x.Id == id);
             if (auction == null) return NotFound();
 
-            // TODO: check seller same as current user
+            // check seller same as current user
+            if (auction.Seller != User.Identity.Name) return Forbid();
+
             auction.Item.Make = auctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = auctionDto.Model ?? auction.Item.Model;
             auction.Item.Color = auctionDto.Color ?? auction.Item.Color;
@@ -99,13 +104,16 @@ namespace AuctionService.Controllers
             return Ok();
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
             if (auction == null) return NotFound();
 
-            // TODO: Check seller == username
+            // check seller same as current user
+            if (auction.Seller != User.Identity.Name) return Forbid();
+
             _context.Auctions.Remove(auction);
             // Publish to Queue (before saving to DB for Atomicity)
             await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
